requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
motor>=3.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic[email]>=2.6.0
pyjwt>=2.8.0
bcrypt>=4.1.2
python-multipart>=0.0.9
pdfminer.six>=20221105
dnspython>=2.6.1
httpx>=0.27.0
selenium>=4.18.0
google-generativeai>=0.5.0
